Clean up and re-order code
